ft_trans/backend/chat/requirements.txt
channels-redis==4.0
channels==4.0
PyJWT==2.7.0
django==4.2
daphne==4.0
redis